+++ src/main/java/com/boozeonwheel/product/repository/master/ProductMasterRespositoryCustom.java
@@ -8,9 +8,9 @@ import com.mongodb.client.result.DeleteResult;
 import com.mongodb.client.result.UpdateResult;
 
 public interface ProductMasterRespositoryCustom {
-	public Master findByProductMasterId(Integer id);
-	public List<Master> findByProductCategoryId(Integer id);
-	public DeleteResult deleteProductMaster(Integer id);
+	public Master findByProductMasterId(long id);
+	public List<Master> findByProductCategoryId(long id);
+	public DeleteResult deleteProductMaster(long id);
 	public void addAllProductMasters(List<Master> masterList);
 	public void deleteAllProductMaster();
 	public UpdateResult updateProductMaster(MasterDTO master);
